Update features for ripgrep v11.0
Adds new ripgrep features noted by BurntSushi in GitHub issue #97, plus
the --binary flag added in the ripgrep v11.0.0 release. In order to
address #95, it expands the "Add filter for ignoring files" section for
both ack and rg in hopes of succinctly explaining one uses arguments to
filter and one loads an .ignore-patterned file but both apply filters.

A typo was fixed.